backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# LangGraph and LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Web3 / Blockchain
web3>=6.11.0
eth-account>=0.10.0

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
pydantic>=2.5.0
python-multipart>=0.0.6
